langchain/day5/requirements.txt
langchain
langchain-community
langchain-ollama
langgraph
python-dotenv
